pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "aspora-tms"
version = "0.1.0"
description = "Aspora Treasury Management System — Operations Agent"
requires-python = ">=3.11"
dependencies = [
    # Core async / settings
    "pydantic>=2.6",
    "pydantic-settings>=2.2",

    # Message bus
    "redis[hiredis]>=5.0",

    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",          # PostgreSQL async driver
    "alembic>=1.13",          # DB migrations

    # Scheduling
    "apscheduler>=3.10",

    # HTTP / API
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",            # Async HTTP for bank APIs

    # Utilities
    "python-dateutil>=2.9",
    "pytz>=2024.1",
    "orjson>=3.10",           # Fast JSON
    "structlog>=24.1",        # Structured logging
    "tenacity>=8.3",          # Retry logic

    # Security
    "cryptography>=42.0",
    "PyJWT>=2.8",

    # Holiday calendars
    "holidays>=0.46",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "freezegun>=1.4",         # Freeze time in tests
    "fakeredis>=2.21",        # In-memory Redis for tests
    "factory-boy>=3.3",       # Test fixtures
    "ruff>=0.4",              # Linter
    "mypy>=1.9",              # Type checking
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
